+++ Editor/treeview/ItemGui/ISelectable.cs
@@ -7,5 +7,6 @@ namespace UnityEditor.Timeline
         void Select();
         bool IsSelected();
         void Deselect();
+        bool CanSelect();
     }
 }